+++ tests/integration_test.rs
@@ -183,6 +183,19 @@ Type `everything --help` or `man 8 the-everything` for more information
 }
 
 #[test]
+fn test_everything_raise_chain() {
+    check(
+        bin_path("examples/everything"),
+        &["--raise-chain", "p1", "p2", "p3"],
+        1,
+        Behavior::Null,
+        Behavior::Inline(
+            "everything: Outermost problem: Intermediate problem: Innermost problem\n".to_owned(),
+        ),
+    );
+}
+
+#[test]
 fn test_everything_help() {
     check(
         bin_path("examples/everything"),
@@ -195,6 +208,7 @@ Options:
     -h, --help          show command-line usage information and exit
         --version       show version information and exit
     -p, --print-args    print free arguments
+        --raise-chain   raises an error with causes
         --raise-error   raises an explicit usage error
 
 Arguments: